Refactor the code to make compiling for more platforms easier.
Each platform now has it's own directory. The files in the platform
directory have all moved to either the msvc or unix directories. Much
of what was in platform.h is now in system.h, and the rest has moved
into the appropriate platform.h file.

See README.md for information on the new structure.