Unix build cleanup as to make multiple platform support easer.
unistd.h is now the unix platform library. The old platform library has
moved to the example directory.

We check for INCLUDE_LICENSE being defined to add the license to the
interpreter.

HAVE_SETJMP is gone, as it's a C89 feature, so if you don't have it we
don't build.

NO_FP (to become NO_MATH) and NO_STDIO now disable the math & stdio
libraries.